requirements.txt
Flask==2.3.2
lxml==4.9.2
#numpy==1.6.1
requests-futures==1.0.1
alembic==1.11.1
bcrypt==4.0.1
cffi==1.15.1
click==8.1.3
dnspython==2.3.0
email-validator==2.0.0.post2
Flask-Bcrypt==1.0.1
Flask-Login==0.6.2
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.1.1
greenlet==2.0.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.3
pycparser==2.21
python-dateutil==2.8.2
python-editor==1.0.4
six==1.16.0
SQLAlchemy==2.0.17
Werkzeug==2.3.6
WTForms==3.0.1
pyEX==0.5.0
alpha_vantage==2.3.1